+++ app/cms/constants/index.ts
@@ -1,3 +1,6 @@
+export const DEFAULT_REPO_OWNER = "onflow"
+export const DEFAULT_CONTENT_PATH = "docs"
+
 export const DEFAULT_SITE_TITLE = "Welcome to the Flow Developer Portal"
 export const DEFAULT_SITE_DESCRIPTION =
   "Your resource for everything you'll need to build in Flow"
