+++ app/cms/constants/repos/display-names.ts
@@ -1,0 +1,12 @@
+/* Overriden display names (defaults to dashes converted to spaces then capitalized) */
+export const displayNames: Partial<Record<string, string>> = {
+  "flow-cli": "Flow CLI",
+  "flow-js-testing": "Flow JS Testing",
+  "flow-go-sdk": "Flow Go SDK",
+  "fcl-js": "Flow Client Library (JS)",
+  "vscode-extension": "VS Code Extension",
+  "dapp-development": "DApp Development",
+  "nft-marketplace": "NFT Marketplace",
+  fusd: "FUSD",
+  faq: "FAQ",
+}
