+++ app/cms/constants/repos/trusted.ts
@@ -1,0 +1,1 @@
+export const trustedRepositories = [""]
